pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deepeval"
version = "0.1.0"
description = "A comprehensive evaluation framework for LLMs"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "DeepEval Team", email = "team@deepeval.com" }
]
dependencies = [
    "python-dotenv>=1.0.0",
    "streamlit>=1.24.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
cloud = [
    "requests>=2.31.0",
    "pydantic>=2.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["deepeval"] 